refactor(tests): move notification test fixtures to conftest.py
