+++ trustpoint/notifications/tests/test_models.py
@@ -1,9 +1,7 @@
 """Tests for the NotificationModel model and related functionality in the notifications app."""
-import typing
 
 import pytest
 from devices.models import DeviceModel
-from django.core.management import call_command
 from django.utils import timezone
 from pki.models import DomainModel, IssuingCaModel
 
@@ -17,54 +15,6 @@ from notifications.models import (
 DEFAULT_EXPIRY_WARNING_DAYS = 30
 DEFAULT_RSA_KEY_SIZE = 2048
 EXPECTED_STATUS_COUNT = 2
-
-@pytest.fixture
-def setup_test_issuing_ca() -> IssuingCaModel :
-    """Use custom management command to create a test Issuing CA."""
-    call_command('create_single_test_issuing_ca')
-    issuing_ca = IssuingCaModel.objects.get(unique_name='issuing-ca-a-test-fixture')
-    return typing.cast(IssuingCaModel, issuing_ca)
-
-
-
-@pytest.fixture
-def test_domain(setup_test_issuing_ca: IssuingCaModel) -> DomainModel:
-    """Create a domain linked to the test issuing CA."""
-    domain, _ = DomainModel.objects.get_or_create(
-        unique_name='test-domain',
-        defaults={'issuing_ca': setup_test_issuing_ca},
-    )
-    domain.issuing_ca = setup_test_issuing_ca
-    domain.save()
-    return domain
-
-@pytest.fixture
-def test_device(test_domain: DomainModel) -> DeviceModel:
-    """Create a test device fixture."""
-    device = DeviceModel.objects.create(
-        common_name='test-device-1',
-        serial_number='TEST123456',
-        domain=test_domain,
-        onboarding_status=DeviceModel.OnboardingStatus.NO_ONBOARDING,
-        onboarding_protocol=DeviceModel.OnboardingProtocol.EST_PASSWORD,
-        pki_protocol=DeviceModel.PkiProtocol.EST_PASSWORD,
-    )
-    return typing.cast(DeviceModel, device)
-
-
-@pytest.fixture
-def test_message() -> NotificationMessageModel:
-    """Create a NotificationMessageModel used for custom notifications."""
-    return NotificationMessageModel.objects.create(
-        short_description='Test short',
-        long_description='Test long description'
-    )
-
-
-@pytest.fixture
-def test_status() -> NotificationStatus:
-    """Create a notification status object."""
-    return NotificationStatus.objects.create(status=NotificationStatus.StatusChoices.NEW)
 
 
 @pytest.mark.django_db
@@ -231,7 +181,7 @@ def test_notification_for_issuing_ca_creation(
         issuing_ca=setup_test_issuing_ca,
         message_data={
             'ca_name': setup_test_issuing_ca.unique_name,
-            'ca_type': setup_test_issuing_ca.IssuingCaTypeChoice(setup_test_issuing_ca.issuing_ca_type).label,
+            'ca_type': str(setup_test_issuing_ca.IssuingCaTypeChoice(setup_test_issuing_ca.issuing_ca_type).label),
             'common_name': setup_test_issuing_ca.common_name
         }
     )
